# packages/api/openapi.yaml
openapi: 3.0.3
info:
  title: Percolator API
  version: 0.1.0
  description: |
    Percolator is a high-performance perpetual futures trading engine built on Solana.
    This API provides read access to market data, trades, funding rates, and platform statistics.
    
    ## Features
    - Real-time market data and pricing
    - Trade history and volume analytics
    - Funding rate calculations and history
    - Open interest tracking
    - Insurance fund monitoring
    - Platform-wide statistics
    - WebSocket support for live updates
    
    ## Rate Limits
    - Read endpoints: 100 requests/minute
    - Write endpoints: 20 requests/minute
    
    ## Caching
    Most endpoints implement response caching with varying TTLs (10-60 seconds) to optimize performance.
    
  contact:
    name: Percolator Team
    url: https://percolator.trade
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001
    description: Local development server
  - url: https://api.percolator.trade
    description: Production server
...
paths:
...
  /markets:
    get:
      tags:
        - Markets
      summary: List all markets
      description: Returns all markets with embedded stats from the markets_with_stats view
      operationId: getMarkets
      responses:
        '200':
          description: Successfully retrieved markets
          content:
            application/json:
              schema:
                type: object
                properties:
                  markets:
                    type: array
                    items:
                      $ref: '#/components/schemas/MarketWithStats'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /markets/stats:
    get:
      tags:
        - Markets
      summary: Get all market stats
      description: Returns statistics for all markets from the database
      operationId: getMarketsStats
      responses:
        '200':
          description: Successfully retrieved market stats
          content:
            application/json:
              schema:
                type: object
                properties:
                  stats:
                    type: array
                    items:
                      $ref: '#/components/schemas/MarketStats'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /funding/global:
    get:
      tags:
        - Funding
      summary: Get current funding rates for all markets
      description: Returns current funding rate data for all markets (60s cache)
      operationId: getGlobalFunding
      responses:
        '200':
          description: Successfully retrieved global funding data
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                  markets:
                    type: array
                    items:
                      $ref: '#/components/schemas/FundingRateSummary'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /stats:
    get:
      tags:
        - Stats
      summary: Get platform statistics
      description: Returns platform-wide aggregated statistics (60s cache)
      operationId: getStats
      responses:
        '200':
          description: Successfully retrieved platform stats
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlatformStats'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
...
  schemas:
...
    MarketWithStats:
      type: object
      properties:
        slabAddress:
          type: string
        mintAddress:
          type: string
        symbol:
          type: string
        name:
          type: string
        decimals:
          type: integer
        deployer:
          type: string
        oracleAuthority:
          type: string
        initialPriceE6:
          type: string
        maxLeverage:
          type: integer
        tradingFeeBps:
          type: integer
        lpCollateral:
          type: string
        matcherContext:
          type: string
        status:
          type: string
        logoUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        totalOpenInterest:
          type: string
          nullable: true
        totalAccounts:
          type: integer
          nullable: true
        lastCrankSlot:
          type: string
          nullable: true
        lastPrice:
          type: string
          nullable: true
        markPrice:
          type: string
          nullable: true
        indexPrice:
          type: string
          nullable: true
        fundingRate:
          type: string
          nullable: true
        netLpPos:
          type: string
          nullable: true
    MarketStats:
      type: object
      properties:
        slab_address:
          type: string
        total_open_interest:
          type: string
        net_lp_pos:
          type: string
        lp_sum_abs:
          type: string
        lp_max_abs:
          type: string
        last_price:
          type: string
        mark_price:
          type: string
        index_price:
          type: string
        funding_rate:
          type: string
        insurance_balance:
          type: string
        insurance_fee_revenue:
          type: string
        volume_24h:
          type: string
        last_crank_slot:
          type: string
        updated_at:
          type: string
          format: date-time
...
    FundingRateSummary:
      type: object
      properties:
        slabAddress:
          type: string
        currentRateBpsPerSlot:
          type: number
          description: Funding rate in basis points per slot
        hourlyRatePercent:
          type: number
          description: Hourly funding rate as percentage
        dailyRatePercent:
          type: number
          description: Daily funding rate as percentage
        netLpPosition:
          type: string
          description: Net LP position (inventory imbalance)
...
    PlatformStats:
      type: object
      properties:
        totalMarkets:
          type: integer
          description: Total number of markets
        volume24h:
          type: string
          description: Aggregate 24h volume across all markets
        totalOpenInterest:
          type: string
          description: Total open interest across all markets
        uniqueDeployers:
          type: integer
          description: Number of unique market deployers
        trades24h:
          type: integer
          description: Total number of trades in last 24 hours
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
        details:
          type: string
          description: Additional error details
        hint:
          type: string
          description: Hint for resolving the error
  responses:
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Failed to fetch data
            details: Database connection timeout